# api/adminapi/openapi.yaml
openapi: 3.0.2
info:
  title: Federation Admin
  version: 0.1.0
  description: >-
    This API enables management of federated entity data and metadata within an OpenID Federation
    environment.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8765
    description: Local development server
...
paths:
...
  /api/v1/admin/entity-configuration:
    get:
      tags:
        - Entity Configuration
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnyValue'
              examples:
                entity_configuration:
                  value:
                    iss: https://example.org
                    sub: https://example.org
                    authority_hints:
                      - https://authority.example.com
                    metadata:
                      federation_entity:
                        federation_list_endpoint: https://example.org/list
                        federation_fetch_endpoint: https://example.org/fetch
                        federation_resolve_endpoint: https://example.org/resolve
                        display_name: Example
                        organization_name: Example Organization
                        contacts:
                          - support@example.org
                    jwks:
                      keys:
                        - kty: EC
                          crv: P-256
                          kid: '12345'
                          x: o-7zraXK...
                          y: Xr_wCzJ1...
                          alg: ES256
                    custom_claim: custom_value
          description: >-
            Successful response returning the entity configuration as it would also be returned from
            the /.well-known/openid-federation endpoint but as JSON.
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: getEntityConfiguration
      summary: Get the entity configuration JSON
...
components:
  schemas:
...
    AnyValue:
      description: Any JSON value. This intentionally permits any type.
      type: object
      additionalProperties: true
      example:
        Example: value
    ErrorResponse:
      title: Root Type for ErrorResponse
      description: Standard error response with code and description.
      required:
        - error
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
      example:
        error: not found
        error_description: resource not found
...
  responses:
...
    ServerError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server error:
              value:
                error: server_error
                error_description: foobar
      description: Internal server error
...